Thangdv20

+++ src/views/apps/cves/CVEScanList.vue
@@ -25,8 +25,8 @@
                             </router-link>
                         </el-tooltip>
 
-                        <el-popconfirm width="300" confirm-button-text="Đồng ý"  cancel-button-text="Không" icon-color="#626AEF"
-                            title="Bạn có chắc muốn scan CVE không?" @confirm="submit">
+                        <el-popconfirm width="300" confirm-button-text="Đồng ý" cancel-button-text="Không"
+                            icon-color="#626AEF" title="Bạn có chắc muốn scan CVE không?" @confirm="submit">
                             <template #reference>
                                 <span>
                                     <el-tooltip class="box-item" effect="dark" hide-after="0" content="Thêm mới"
@@ -47,9 +47,9 @@
                             <span class="me-2">{{ selectedIds.length }}</span>Selected
                         </div>
                         <el-tooltip class="box-item" effect="dark" hide-after="0" content="Xóa" placement="top">
-                            <button type="button" data-bs-target="#kt_modal_delete" data-bs-toggle="modal"
-                                class="btn btn-danger btn-sm " :disabled="disabled">
-                                Delete Selected
+                            <button type="button" @click="deleteSelectd()" class="btn btn-danger btn-sm "
+                                :disabled="disabled">
+                                Xóa mục đã chọn
                             </button>
                         </el-tooltip>
                     </div>
@@ -122,9 +122,9 @@
                                                     </el-progress>
                                                 </div>
                                                 <span class="fd-inline-block text-center fs-7 w-100">
-                                                        <span class="badge badge-light-primary">Note : </span>
-                                                        Tiến trình CVE
-                                                    </span>
+                                                    <span class="badge badge-light-primary">Note : </span>
+                                                    Tiến trình CVE
+                                                </span>
                                             </div>
                                             <div class="col-8">
                                                 <div class="row  ">
@@ -198,36 +198,6 @@
             </div>
         </div>
     </div>
-
-    <div class="modal fade" tabindex="-1" id="kt_modal_delete" ref="ModalDelete" aria-hidden="true">
-        <div class="modal-dialog modal-dialog-centered">
-            <div class="modal-content">
-
-                <div class="modal-header">
-                    <h3 class="modal-title">Xác nhận xóa scan</h3>
-
-                    <div class="btn btn-icon btn-sm btn-active-light-primary ms-2" data-bs-dismiss="modal"
-                        aria-label="Close">
-                        <span class="svg-icon svg-icon-2x"></span>
-                    </div>
-                </div>
-                <div class="modal-body">
-                    <p style="font-size: 16px;">Bạn có chắc chắn muốn xóa <span v-if="selectedIds.length > 0"
-                            class="fw-bold" style="color:red;">{{ selectedIds.length
-                            }}</span> bản ghi này không?
-                    </p>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-sm  btn-light" data-bs-dismiss="modal">
-                        Hủy bỏ
-                    </button>
-                    <button type="button" class="btn btn-sm  btn-primary" @click.passive="deleteFewSubscriptions()">
-                        Đồng ý
-                    </button>
-                </div>
-            </div>
-        </div>
-    </div>
 </template>
 
 <script lang="ts">
@@ -242,7 +212,9 @@ import { hideModal } from "@/core/helpers/dom";
 import { ErrorMessage, Field, Form as VForm } from "vee-validate";
 import { vue3Debounce } from 'vue-debounce';
 import Fillter from "@/views/apps/targets/filterTargetScan.vue";
-
+import { ElMessage, ElMessageBox } from 'element-plus'
+import { Delete } from '@element-plus/icons-vue'
+import { markRaw } from 'vue'
 import Swal from "sweetalert2/dist/sweetalert2.js";
 
 import { Modal } from "bootstrap";
@@ -458,12 +430,26 @@ export default defineComponent({
         });
 
         const selectedIds = ref<Array<number>>([]);
-        const deleteFewSubscriptions = () => {
-            deleteSubscription(selectedIds.value);
+        const deleteSelectd = () => {
+            ElMessageBox.confirm(
+                'Tập tin sẽ được xóa vĩnh viễn. Tiếp tục?',
+                'Xác nhận xóa',
+                {
+                    confirmButtonText: 'Đồng ý',
+                    cancelButtonText: 'Hủy bỏ',
+                    type: 'warning',
+                    icon: markRaw(Delete)
+                }
+            )
+                .then(() => {
+                    deleteSubscription(selectedIds.value);
+                })
+                .catch(() => {
+
+                })
         };
         const disabled = ref<boolean>(false);
 
-        const ModalDelete = ref<null | HTMLElement>(null);
         const deleteSubscription = (ids: Array<number>) => {
             let formData = {
                 'id': ids
@@ -534,7 +520,7 @@ export default defineComponent({
                 },
             }).then(() => {
                 hideModal(newTargetGroupModalRef.value);
-                hideModal(ModalDelete.value);
+
                 hideModal(ModalDetail.value);
             });
         }
@@ -599,8 +585,8 @@ export default defineComponent({
             sort,
             onItemSelect,
             selectedIds,
-            deleteFewSubscriptions,
-            deleteSubscription,
+            deleteSelectd,
+            
             getAssetPath,
 
             // validate
@@ -615,7 +601,7 @@ export default defineComponent({
             newTargetGroupModalRef,
             handleClick,
             errors,
-            ModalDelete,
+
             discardButtonRef,
             // detials
             ModalDetail,
