+++ src/views/apps/targets/TargetReconList.vue
@@ -29,9 +29,8 @@
                             <span class="me-2">{{ selectedIds.length }}</span>Selected
                         </div>
                         <el-tooltip class="box-item" effect="dark" hide-after="0" content="Xóa" placement="top">
-                            <button type="button" data-bs-target="#kt_modal_delete" data-bs-toggle="modal"
-                                class="btn btn-danger btn-sm ">
-                                Delete Selected
+                            <button type="button" @click="deleteSelectd()" class="btn btn-danger btn-sm ">
+                                Xóa mục đã chọn
                             </button>
                         </el-tooltip>
                     </div>
@@ -114,37 +113,6 @@
                         </button>
                     </div>
                 </VForm>
-            </div>
-        </div>
-    </div>
-
-    <div class="modal fade" tabindex="-1" id="kt_modal_delete" ref="ModalDelete" aria-hidden="true">
-        <div class="modal-dialog modal-dialog-centered">
-            <div class="modal-content">
-
-                <div class="modal-header">
-                    <h3 class="modal-title">Xác nhận xóa Recon</h3>
-
-                    <div class="btn btn-icon btn-sm btn-active-light-primary ms-2" data-bs-dismiss="modal"
-                        aria-label="Close">
-                        <span class="svg-icon svg-icon-2x"></span>
-                    </div>
-                </div>
-                <div class="modal-body">
-                    <p style="font-size: 16px;">Bạn có chắc chắn muốn xóa <span v-if="selectedIds.length > 0"
-                            class="fw-bold" style="color:red;">{{ selectedIds.length
-                            }}</span> bản ghi này không?
-                    </p>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-sm  btn-light" :disabled="disabled" data-bs-dismiss="modal">
-                        Hủy bỏ
-                    </button>
-                    <button type="button" class="btn btn-sm  btn-primary" :disabled="disabled"
-                        @click.passive="deleteFewSubscriptions()">
-                        Đồng ý
-                    </button>
-                </div>
             </div>
         </div>
     </div>
@@ -234,8 +202,9 @@ import Swal from "sweetalert2/dist/sweetalert2.js";
 
 import { Modal } from "bootstrap";
 import { useRoute } from 'vue-router';
-import { ElMessage } from 'element-plus'
-
+import { ElMessage, ElMessageBox } from 'element-plus'
+import { Delete } from '@element-plus/icons-vue'
+import { markRaw } from 'vue'
 interface APIData {
     status: string;
     created_at: string;
@@ -883,7 +852,7 @@ export default defineComponent({
         const route = useRoute();
         const reconID = ref<null | number | any>(route.params.id ?? '');
 
-        const getData = async  () => {
+        const getData = async () => {
             loading.value = true;
             return ApiService.get(`/recon/${reconID.value}/target?search_recon=${query.value}&search_status=${filterStatus.value}&page=${currentPage.value}&page_size=${itemsPerPage.value}&ordering=${orderingID.value}`)
                 .then(({ data }) => {
@@ -900,7 +869,7 @@ export default defineComponent({
 
         // tính thời gian
         const eventTime = ref<number | any>('30000');
-        let intervalId : any;
+        let intervalId: any;
         const startTimer = () => {
             intervalId = setInterval(() => {
                 getData();
@@ -920,12 +889,27 @@ export default defineComponent({
         });
 
         const selectedIds = ref<Array<number>>([]);
-        const deleteFewSubscriptions = () => {
-            deleteSubscription(selectedIds.value); 
+        const deleteSelectd = () => {
+            ElMessageBox.confirm(
+                'Tập tin sẽ được xóa vĩnh viễn. Tiếp tục?',
+                'Xác nhận xóa',
+                {
+                    confirmButtonText: 'Đồng ý',
+                    cancelButtonText: 'Hủy bỏ',
+                    type: 'warning',
+                    icon: markRaw(Delete)
+                }
+            )
+                .then(() => {
+                    deleteSubscription(selectedIds.value);
+                })
+                .catch(() => {
+
+                })
         };
         const disabled = ref<boolean>(false);
 
-        const ModalDelete = ref<null | HTMLElement>(null);
+
         const deleteSubscription = (ids: Array<number>) => {
             if (ids) {
                 disabled.value = true
@@ -1051,7 +1035,7 @@ export default defineComponent({
                 },
             }).then(() => {
                 hideModal(newTargetGroupModalRef.value);
-                hideModal(ModalDelete.value);
+
                 hideModal(ModalDetail.value);
             });
         }
@@ -1155,8 +1139,8 @@ export default defineComponent({
             sort,
             onItemSelect,
             selectedIds,
-            deleteFewSubscriptions,
-            deleteSubscription,
+            deleteSelectd,
+            
             getAssetPath,
 
             // validate
@@ -1173,7 +1157,7 @@ export default defineComponent({
             newTargetGroupModalRef,
             handleClick,
             errors,
-            ModalDelete,
+
             discardButtonRef,
             clearHeaderOptions,
             // detials
